backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
Django==6.0.2
django-ninja==1.5.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
django-cors-headers
django-ninja-jwt==5.3.1
google-auth==2.28.1
requests==2.31.0
email-validator
